requirements.txt
google-generativeai
pygame~=2.6.1
PySimpleGUI~=5.0.9
configparser~=7.2.0
protobuf~=5.29.2
defusedxml~=0.7.1
ipython~=8.30.0
PyQt6~=6.6.1
numpy~=1.26.4
pillow~=11.1.0
olefile~=0.47
bleach~=6.2.0
lxml~=5.3.1
beautifulsoup4~=4.13.3
soupsieve~=2.6
pip~=25.0.1
attrs~=25.1.0
distro~=1.9.0
wheel~=0.45.1
PySocks~=1.7.1
pyOpenSSL~=25.0.0
cryptography~=41.0.7
filelock~=3.17.0
redis~=4.3.4
rsa~=4.9
pyasn1~=0.6.1
grpcio~=1.69.0
requests~=2.32.3
rich~=13.9.4
colorama~=0.4.6
charset-normalizer~=3.3.2
setuptools~=75.8.0
pytz~=2025.1
cffi~=1.17.1
pyinstaller~=6.11.1
proto-plus~=1.25.0
aiohttp~=3.10.11
pyasn1_modules~=0.4.1
cachetools~=5.5.0
certifi~=2025.1.31
urllib3~=2.3.0
packaging~=24.2
google-api-python-client~=2.157.0
grpcio-status~=1.69.0
typing_extensions~=4.12.2
pydantic~=2.10.6
pandas~=2.2.3+dfsg
httplib2~=0.22.0
tqdm~=4.67.1
zstandard~=0.23.0
h2~=4.1.0
six~=1.17.0
python-dateutil~=2.9.0
pyparsing~=3.1.2
email_validator~=2.2.0
pydantic_core~=2.27.2
Deprecated~=1.2.18
annotated-types~=0.7.0
matplotlib~=3.8.3
idna~=3.10
chardet~=5.2.0
simplejson~=3.19.3
contourpy~=1.3.1
fonttools~=4.55.3
sympy~=1.13.3
ufoLib2~=0.17.0
lz4~=4.4.0+dfsg
fs~=2.4.16
pycairo~=1.26.1
scipy~=1.14.1
unicodedata2~=15.1.0
Jinja2~=3.1.5
cycler~=0.12.1
PyGObject~=3.50.0
tornado~=6.4.2
kiwisolver~=1.4.7
autocommand~=2.2.2
inflect~=7.3.1
more-itertools~=10.6.0
typeguard~=4.4.2
platformdirs~=4.3.6
zipp~=3.21.0
Pygments~=2.18.0
uritemplate~=4.1.1
google-auth-httplib2~=0.2.0